+++ tests/test_metrics/test_losses.py
@@ -82,7 +82,7 @@ def test_hvu_loss():
 def test_cross_entropy_loss():
     cls_scores = torch.rand((3, 4))
     sparse_gt_labels = torch.LongTensor([0, 1, 2]).squeeze()
-    non_sparse_gt_labels = torch.FloatTensor([[1, 0, 0, 0], [0, 1, 0, 0],
+    soft_gt_labels = torch.FloatTensor([[1, 0, 0, 0], [0, 1, 0, 0],
                                               [0, 0, 1, 0]]).squeeze()
 
     # sparse label without weight
@@ -100,18 +100,18 @@ def test_cross_entropy_loss():
         output_loss,
         F.cross_entropy(cls_scores, sparse_gt_labels, weight=weight))
 
-    # non-sparse label without class weight
+    # soft label without class weight
     cross_entropy_loss = CrossEntropyLoss()
-    output_loss = cross_entropy_loss(cls_scores, non_sparse_gt_labels)
+    output_loss = cross_entropy_loss(cls_scores, soft_gt_labels)
 
     assert_almost_equal(
         output_loss.numpy(),
         F.cross_entropy(cls_scores, sparse_gt_labels).numpy(),
         decimal=4)
 
-    # non-sparse label with class weight
+    # soft label with class weight
     cross_entropy_loss = CrossEntropyLoss(class_weight=class_weight)
-    output_loss = cross_entropy_loss(cls_scores, non_sparse_gt_labels)
+    output_loss = cross_entropy_loss(cls_scores, soft_gt_labels)
     assert_almost_equal(
         output_loss.numpy(),
         F.cross_entropy(cls_scores, sparse_gt_labels).numpy(),